chore: Move new list component
